panel/requirements.txt
# GIFT Oceanographic Intelligence Platform - Dependencies
# Python 3.9+

# ===== CORE FRAMEWORK =====
streamlit==1.29.0
streamlit-option-menu==0.3.6

# ===== DATA MANIPULATION =====
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# ===== VISUALIZATION =====
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
folium==0.15.1
streamlit-folium==0.15.1

# ===== STATISTICS & ML =====
scikit-learn==1.3.2
statsmodels==0.14.0

# ===== GEOSPATIAL =====
geopandas==0.14.1
shapely==2.0.2

# ===== DATABASE (Optional for future) =====
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9

# ===== UTILITIES =====
python-dotenv==1.0.0
pyyaml==6.0.1
